Feature/setup wizard

+++ controllers/workspace.controller.js
@@ -4,6 +4,7 @@ const CheckRole=require("../middlewares/CheckRoleMiddleware")
 const User=require("../models/user.model")
 const Joi = require("joi");
 const createWorkspace = async (req, res,next) => {
+  if (!CheckRole(req, res, ["admin"])) return;
   try {
  
     const { error, value } = workspaceValidation.validate(req.body, {
@@ -50,8 +51,9 @@ const createWorkspace = async (req, res,next) => {
 
 
 const getWorkspaces = async (req, res, next) => {
+  if (!CheckRole(req, res, ["admin"])) return;
   try {
-    const workspaces = await Workspace.find({ ownerId: req.user._id }).populate("ownerId").sort({
+    const workspaces = await Workspace.find({ ownerId: req.user.id }).populate("ownerId").sort({
       createdAt: -1,
     });
 
@@ -64,9 +66,10 @@ const getWorkspaces = async (req, res, next) => {
 
 
 const getWorkspace = async (req, res, next) => {
+  if (!CheckRole(req, res, ["admin"])) return;
   try {
     const targetWorkspace = await Workspace.findOne({
-      _id: req.params.id,
+      _id: req.params.workspaceId,
       ownerId: req.user.id,
     }).populate("ownerId");
 
@@ -87,6 +90,7 @@ const getWorkspace = async (req, res, next) => {
 
 
 const updateWorkspace = async (req, res, next) => {
+  if (!CheckRole(req, res, ["admin"])) return;
   try {
    
     const { error, value } = workspaceValidation.validate(req.body, {
@@ -114,7 +118,7 @@ const updateWorkspace = async (req, res, next) => {
 
     const targetWorkspace = await Workspace.findOneAndUpdate(
       {
-        _id: req.params.id,
+        _id: req.params.workspaceId,
         ownerId: req.user.id,
       },
       value,
@@ -142,10 +146,11 @@ const updateWorkspace = async (req, res, next) => {
 
 
 const deleteWorkspace = async (req, res, next) => {
+  if (!CheckRole(req, res, ["admin"])) return;
   try {
 
     const targetWorkspace = await Workspace.findOneAndDelete({
-      _id: req.params.id,
+      _id: req.params.workspaceId,
       ownerId: req.user.id,
     });
